+++ saleor/graphql/account/types.py
@@ -10,6 +10,7 @@ from promise import Promise
 from ...account import models
 from ...checkout.utils import get_user_checkout
 from ...core.exceptions import PermissionDenied
+from ...graphql.meta.inputs import MetadataInput
 from ...order import OrderStatus
 from ...permission.auth_filters import AuthorizationFilters
 from ...permission.enums import AccountPermissions, AppPermission, OrderPermissions
@@ -83,6 +84,17 @@ class AddressInput(BaseInputObjectType):
             "Phone numbers are validated with Google's "
             "[libphonenumber](https://github.com/google/libphonenumber) library."
         )
+    )
+
+    metadata = graphene.List(
+        graphene.NonNull(MetadataInput),
+        description="Checkout public metadata.",
+        required=False,
+    )
+    private_metadata = graphene.List(
+        graphene.NonNull(MetadataInput),
+        description="Checkout private metadata.",
+        required=False,
     )