+++ CHANGELOG.md
@@ -69,6 +69,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add `search` to `giftCards` query - #13173 by @zedzior
 - Add `ProductBulkTranslate` mutation - #13329 by @SzymJ
 - Add `ProductVariantBulkTranslate` mutation - #13329 by @SzymJ
+- Add `metadata` and `privateMetadata` to `AddressInput` field - #13248 by @Air-t
 
 ### Saleor Apps